chore: move to /utils/graphql
