+++ server/utils/graphql/validators.py
@@ -1,5 +1,5 @@
 from typing import Dict, Any
-from utils.graphql import execute_query
+from .common import execute_query
 
 
 def get_graphql_validators(chain: str, network: str) -> Dict[str, Any]:
